split parton and pseudo top
